taint/test/mochitest/test_dom.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Check DOM Element Sources</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted " + str);
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted " + str);
      }

      function check_taint_source(str, src) {
          var flow = str.taint[0].flow;
          var op = flow[flow.length-1];
          ok(op.operation == src, "Source operation check: '" + op.operation + "' =? '" + src + "'");
      }

      add_task(async function test_untainted() {
          check_untainted("hello");
      });

      add_task(async function test_tainted() {
          check_tainted(String.tainted("hello"));
      });

      add_task(async function test_div() {
          check_tainted(document.getElementById("content").getAttribute("test"));
          check_tainted(document.getElementById("content").getAttribute("long"));
      });

      add_task(async function test_div_dynamic() {
          var d = document.createElement("div");
          d.setAttribute("test", "hello");
          check_tainted(d.getAttribute("test"));
      });

      add_task(async function test_anchor_href() {
          check_tainted(document.getElementsByTagName("a")[0].href);
          check_tainted(document.getElementsByTagName("a")[0].getAttribute("href"));
      });

      add_task(async function test_img_src() {
          check_tainted(document.getElementsByTagName("img")[0].src);
          check_tainted(document.getElementsByTagName("img")[0].getAttribute("src"));
      });

      add_task(async function test_img_alt() {
          check_tainted(document.getElementsByTagName("img")[0].alt);
          check_tainted(document.getElementsByTagName("img")[0].getAttribute("alt"));
      });

      add_task(async function test_script_src() {
          check_tainted(document.getElementById("s1").src);
          check_tainted(document.getElementById("s1").getAttribute("src"));
      });

      add_task(async function test_element_no_selector() {
          var d = document.createElement("div");
          d.setAttribute("test", "hello");
          var tainted = d.getAttribute("test");

          check_tainted(tainted);
          check_taint_source(tainted, "element.attribute");
      });

      add_task(async function test_element_no_selector_tainted() {
          var d = document.createElement("div");
          d.setAttribute("test", String.tainted("hello"));
          var tainted = d.getAttribute("test");

          check_tainted(tainted);
          check_taint_source(tainted, "manual taint source");
      });

      add_task(async function test_div_innerhtml_tainted() {
          var d = document.createElement("div");
          d.innerHTML = String.tainted("<div test='helllo'>Content Here</div>");
          var tainted = d.children[0].getAttribute("test");

          check_tainted(tainted);
          check_taint_source(tainted, "manual taint source");
      });

      add_task(async function test_div_by_id() {
          var element = document.getElementById("content_by_id");
          var tainted = element.getAttribute("test");

          check_tainted(tainted);
          check_taint_source(tainted, "document.getElementById");
      });


      add_task(async function test_div_by_id_already_tainted() {
          var selector = String.tainted("content_by_id_already_tainted")
          var element = document.getElementById(selector);
          var tainted = element.getAttribute("test");

          check_tainted(tainted);
          check_taint_source(tainted, "manual taint source");
      });

    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none" test="hello" long="This should be so long that it does not get atomized!"></div>
    <pre id="test"></pre>
    <a href="https://example.com">Visit Example Website</a>
    <img src="default16.png" alt="A test file">
    <script id="s1" src="xx"></script>

    <div id="content_by_id" style="display: none" test="hello" long="This should be so long that it does not get atomized!"></div>
    <div id="content_by_id_already_tainted" style="display: none" test="hello" long="This should be so long that it does not get atomized!"></div>

  </body>
</html>
